+++ app/src/main/java/dev/yokai/core/di/DomainModule.kt
@@ -2,6 +2,7 @@ package dev.yokai.core.di
 
 import dev.yokai.domain.extension.repo.ExtensionRepoRepository
 import dev.yokai.data.extension.repo.ExtensionRepoRepositoryImpl
+import dev.yokai.data.library.custom.CustomMangaRepositoryImpl
 import dev.yokai.domain.extension.interactor.TrustExtension
 import dev.yokai.domain.extension.repo.interactor.CreateExtensionRepo
 import dev.yokai.domain.extension.repo.interactor.DeleteExtensionRepo
@@ -9,6 +10,10 @@ import dev.yokai.domain.extension.repo.interactor.GetExtensionRepo
 import dev.yokai.domain.extension.repo.interactor.GetExtensionRepoCount
 import dev.yokai.domain.extension.repo.interactor.ReplaceExtensionRepo
 import dev.yokai.domain.extension.repo.interactor.UpdateExtensionRepo
+import dev.yokai.domain.library.custom.CustomMangaRepository
+import dev.yokai.domain.library.custom.interactor.CreateCustomManga
+import dev.yokai.domain.library.custom.interactor.DeleteCustomManga
+import dev.yokai.domain.library.custom.interactor.GetCustomManga
 import uy.kohesive.injekt.api.InjektModule
 import uy.kohesive.injekt.api.InjektRegistrar
 import uy.kohesive.injekt.api.addFactory
@@ -17,6 +22,8 @@ import uy.kohesive.injekt.api.get
 
 class DomainModule : InjektModule {
     override fun InjektRegistrar.registerInjectables() {
+        addFactory { TrustExtension(get(), get()) }
+
         addSingletonFactory<ExtensionRepoRepository> { ExtensionRepoRepositoryImpl(get()) }
         addFactory { CreateExtensionRepo(get()) }
         addFactory { DeleteExtensionRepo(get()) }
@@ -25,6 +32,9 @@ class DomainModule : InjektModule {
         addFactory { ReplaceExtensionRepo(get()) }
         addFactory { UpdateExtensionRepo(get(), get()) }
 
-        addFactory { TrustExtension(get(), get()) }
+        addSingletonFactory<CustomMangaRepository> { CustomMangaRepositoryImpl(get()) }
+        addFactory { CreateCustomManga(get()) }
+        addFactory { DeleteCustomManga(get()) }
+        addFactory { GetCustomManga(get()) }
     }
 }