+++ app/src/main/java/dev/yokai/domain/library/custom/interactor/GetCustomManga.kt
@@ -1,0 +1,11 @@
+package dev.yokai.domain.library.custom.interactor
+
+import dev.yokai.domain.library.custom.CustomMangaRepository
+
+class GetCustomManga(
+    private val customMangaRepository: CustomMangaRepository,
+) {
+    fun subscribeAll() = customMangaRepository.subscribeAll()
+
+    suspend fun getAll() = customMangaRepository.getAll()
+}