+++ app/src/main/java/eu/kanade/tachiyomi/data/library/CustomMangaManager.kt
@@ -6,9 +6,16 @@ import dev.yokai.core.metadata.COMIC_INFO_EDITS_FILE
 import dev.yokai.core.metadata.ComicInfo
 import dev.yokai.core.metadata.ComicInfoPublishingStatus
 import dev.yokai.core.metadata.copyFromComicInfo
+import dev.yokai.domain.library.custom.interactor.CreateCustomManga
+import dev.yokai.domain.library.custom.interactor.DeleteCustomManga
+import dev.yokai.domain.library.custom.interactor.GetCustomManga
+import dev.yokai.domain.library.custom.model.CustomMangaInfo
+import dev.yokai.domain.library.custom.model.CustomMangaInfo.Companion.getMangaInfo
 import eu.kanade.tachiyomi.data.database.models.Manga
 import eu.kanade.tachiyomi.data.database.models.MangaImpl
-import eu.kanade.tachiyomi.util.system.writeText
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromStream
@@ -17,44 +24,46 @@ import nl.adaptivity.xmlutil.serialization.XML
 import nl.adaptivity.xmlutil.serialization.XmlSerialName
 import nl.adaptivity.xmlutil.serialization.XmlValue
 import uy.kohesive.injekt.injectLazy
-import java.io.InputStream
 import java.nio.charset.StandardCharsets
 
 class CustomMangaManager(val context: Context) {
+    private val scope = CoroutineScope(Dispatchers.IO)
+
     private val xml: XML by injectLazy()
 
     private val externalDir = UniFile.fromFile(context.getExternalFilesDir(null))
 
+    private var removedCustomManga = mutableListOf<Long>()
     private var customMangaMap = mutableMapOf<Long, Manga>()
 
+    private val createCustomManga: CreateCustomManga by injectLazy()
+    private val deleteCustomManga: DeleteCustomManga by injectLazy()
+    private val getCustomManga: GetCustomManga by injectLazy()
+
     init {
-        fetchCustomData()
+        scope.launch {
+            fetchCustomData()
+        }
     }
 
     companion object {
         const val EDIT_JSON_FILE = "edits.json"
-
-        fun Manga.toComicInfo(): ComicList.ComicInfoYokai {
-            return ComicList.ComicInfoYokai.create(
-                id = id!!,
-                title = title,
-                author = author,
-                artist = artist,
-                description = description,
-                genre = genre.orEmpty(),
-                status = status,
-            )
-        }
     }
 
     fun getManga(manga: Manga): Manga? = customMangaMap[manga.id]
 
-    private fun fetchCustomData() {
+    private suspend fun fetchCustomData() {
         val comicInfoEdits = externalDir?.findFile(COMIC_INFO_EDITS_FILE)
         val editJson = externalDir?.findFile(EDIT_JSON_FILE)
 
+        val dbMangaInfo = getCustomManga.getAll()
+        customMangaMap = dbMangaInfo.associate { info ->
+            val id = info.mangaId
+            id to info.toManga()
+        }.toMutableMap()
+
         if (comicInfoEdits != null && comicInfoEdits.exists() && comicInfoEdits.isFile) {
-            fetchFromComicInfo(comicInfoEdits.openInputStream())
+            fetchFromComicInfo(comicInfoEdits)
             return
         }
 
@@ -65,20 +74,22 @@ class CustomMangaManager(val context: Context) {
         }
     }
 
-    private fun fetchFromComicInfo(stream: InputStream) {
-        val comicInfoEdits = AndroidXmlReader(stream, StandardCharsets.UTF_8.name()).use {
+    private suspend fun fetchFromComicInfo(comicInfoFile: UniFile) {
+        val comicInfoEdits = AndroidXmlReader(comicInfoFile.openInputStream(), StandardCharsets.UTF_8.name()).use {
             xml.decodeFromReader<ComicList>(it)
         }
 
         if (comicInfoEdits.comics == null) return
 
-        customMangaMap = comicInfoEdits.comics.mapNotNull { obj ->
+        comicInfoEdits.comics.mapNotNull { obj ->
             val id = obj.id ?: return@mapNotNull null
-            id to mangaFromComicInfoObject(id, obj.value)
-        }.toMap().toMutableMap()
+            customMangaMap[id] = mangaFromComicInfoObject(id, obj.value)
+        }
+
+        saveCustomInfo { comicInfoFile.delete() }
     }
 
-    private fun fetchFromLegacyJson(jsonFile: UniFile) {
+    private suspend fun fetchFromLegacyJson(jsonFile: UniFile) {
         val json = try {
             Json.decodeFromStream<MangaList>(jsonFile.openInputStream())
         } catch (e: Exception) {
@@ -86,36 +97,41 @@ class CustomMangaManager(val context: Context) {
         } ?: return
 
         val mangasJson = json.mangas ?: return
-        customMangaMap = mangasJson.mapNotNull { mangaObject ->
+        mangasJson.mapNotNull { mangaObject ->
             val id = mangaObject.id ?: return@mapNotNull null
-            id to mangaObject.toManga()
-        }.toMap().toMutableMap()
+            customMangaMap[id] = mangaObject.toManga()
+        }
 
         saveCustomInfo { jsonFile.delete() }
     }
 
-    fun saveMangaInfo(manga: ComicList.ComicInfoYokai) {
-        val mangaId = manga.id ?: return
-        if (manga.value.series == null &&
-            manga.value.writer == null &&
-            manga.value.penciller == null &&
-            manga.value.summary == null &&
-            manga.value.genre == null &&
-            (manga.value.publishingStatus?.value ?: "Invalid") == "Invalid"
+    suspend fun saveMangaInfo(manga: CustomMangaInfo) {
+        val mangaId = manga.mangaId ?: return
+        if (manga.title == null &&
+            manga.author == null &&
+            manga.artist == null &&
+            manga.description == null &&
+            manga.genre == null &&
+            (manga.status ?: -1) == -1
         ) {
-            customMangaMap.remove(mangaId)
+            customMangaMap[mangaId]?.let {
+                removedCustomManga.add(mangaId)
+                customMangaMap.remove(mangaId)
+            }
         } else {
-            customMangaMap[mangaId] = mangaFromComicInfoObject(mangaId, manga.value)
+            customMangaMap[mangaId] = manga.toManga()
         }
         saveCustomInfo()
     }
 
-    private fun saveCustomInfo(onComplete: () -> Unit = {}) {
-        val comicInfoEdits = externalDir?.createFile(COMIC_INFO_EDITS_FILE) ?: return
+    private suspend fun saveCustomInfo(onComplete: () -> Unit = {}) {
+        deleteCustomManga.bulk(removedCustomManga)
+        removedCustomManga = mutableListOf()
 
-        val edits = customMangaMap.values.map { it.toComicInfo() }
-        if (edits.isNotEmpty() && comicInfoEdits.exists()) {
-            comicInfoEdits.writeText(xml.encodeToString(ComicList.serializer(), ComicList(edits)), onComplete = onComplete)
+        val edits = customMangaMap.values.map { it.getMangaInfo() }
+        if (edits.isNotEmpty()) {
+            createCustomManga.bulk(edits)
+            onComplete()
         }
     }